move status code to separate enums folder
