+++ src/middleware/static_serve.cpp
@@ -32,7 +32,7 @@ bool expresso::middleware::StaticServe::use(expresso::core::Request &req,
 
     res.send(this->getFolderHTML(req, tempPath));
     res.set("Content-Type", "text/html");
-    res.status(expresso::core::STATUS_CODE::OK);
+    res.status(expresso::enums::STATUS_CODE::OK);
     return false;
   }
